+++ flutter_network_layer_core/lib/src/request/request_command.dart
@@ -1,0 +1,37 @@
+import 'package:flutter_network_layer_core/flutter_network_layer_core.dart';
+
+/// The callback for the progress of the request. Can be used to trace uploading
+/// or downloading data.
+typedef OnProgressCallback = void Function(int done, int total);
+
+/// The interface/class to configure a request type.
+///
+/// An implementation should provide the necessary information about the request
+/// such as the path, method, payload type, payload data, headers, progress
+/// callbacks, etc.
+abstract class RequestCommand<T extends ResponseModel> {
+  /// The path of the request.
+  String get path;
+
+  /// The sample model to deserialize the response via its fromJson function.
+  T get sampleModel;
+
+  /// The method of the request. GET, POST, PUT, DELETE, etc.
+  HttpRequestMethod get method => HttpRequestMethod.get;
+
+  /// The type of the payload. Json, form data, etc.
+  RequestPayloadType get payloadType => RequestPayloadType.json;
+
+  /// The payload data of the request. Applicable in the form of Json or form
+  /// data.
+  Map<String, dynamic> get data => const {};
+
+  /// The headers of the request.
+  Map<String, dynamic> get headers => const {};
+
+  /// The callback for the progress of the request.
+  OnProgressCallback? get onSendProgressUpdate => null;
+
+  /// The callback for the progress of the response.
+  OnProgressCallback? get onReceiveProgressUpdate => null;
+}