+++ flutter_network_layer_core/test/utils/test_response_samples.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_network_layer_core/flutter_network_layer_core.dart';
 
-final class ResponseTest1 implements IResponseModel {
+final class ResponseTest1 extends JsonResponseModel {
   const ResponseTest1({required this.field1});
 
   const ResponseTest1.empty() : field1 = '';
@@ -8,7 +8,7 @@ final class ResponseTest1 implements IResponseModel {
   final String field1;
 
   @override
-  IResponseModel fromJson(dynamic json) {
+  ResponseTest1 fromJson(dynamic json) {
     if (json is! Map<String, dynamic>) {
       throw Exception('Invalid JSON: $json');
     }