+++ flutter_network_layer_core/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* `IRequestCommand` interface turned into `RequestCommand` abstract class.
+* `IResponseModel` interface turned into `JsonResponseModel` and `CustomResponseModel` abstract classes.
+* In product level, requests and responses have to be created by extending these abstract classes.
+
 ## 0.1.1
 
 * Readme updated.